+++ NEWS.md
@@ -6,6 +6,9 @@
   more-precise detection of files that changed less than a second after
   the previous backup.
 
+* Changed `conserve tree size` and `conserve source size` to report in MB, not
+  bytes, as used elsewhere in the UI.
+
 ## Conserve 0.6.1 2020-01-25
 
 * Improved performance on incremental backup, by only opening source files if